+++ apps/api/src/events/dto/update-event.dto.ts
@@ -1,0 +1,30 @@
+import { EventType } from '@prisma/client';
+import {
+  IsEnum,
+  IsOptional,
+  IsString,
+  IsUrl,
+  MaxLength,
+  MinLength,
+} from 'class-validator';
+
+export class UpdateEventDto {
+  @IsString()
+  @MinLength(1)
+  @MaxLength(100)
+  @IsOptional()
+  name?: string;
+
+  @IsEnum(EventType)
+  @IsOptional()
+  type?: EventType;
+
+  @IsString()
+  @MaxLength(500)
+  @IsOptional()
+  description?: string;
+
+  @IsUrl()
+  @IsOptional()
+  coverImageUrl?: string;
+}